// src/types/Settings.d.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Aliases } from './Alias';

export type Settings = {
    enable: boolean;
    streaming_only: boolean;
    streaming_mc_only: boolean;
    mc_client: string;
    custom_logs_path: string;
    twitch_bot: string;
    bot_prefix: string;
    custom_bot_id: string;
    disable_hardware_acceleration: boolean;
    start_with_windows: boolean;
    language: string;
    scenes_enable: boolean;
    scenes_name: string;
    scenes_delay: number;
    scenes_hide_in_lobby: boolean;
    scenes_toggle_after_game_end: boolean;
    scenes_auto_obs_config: boolean;
    scenes_obs_ws_port: number;
    scenes_obs_ws_password: string;
    scenes_obs_config_path: string;
    pack_enable: boolean;
    pack_command: string;
    pack_command_text: string;
    pack_hide_overlay: boolean;
    pack_announcements_enable: boolean;
    pack_announcements_text: string;
    server_enable: boolean;
    server_command: string;
    server_command_text: string;
    server_announcements_enable: boolean;
    server_announcements_text: string;
    music_enable: boolean;
    music_ignore_twitch: boolean;
    music_command: string;
    music_command_text: string;
    music_announcements_enable: boolean;
    music_announce_text: string;
    aliases: Aliases;
};